+++ scripts/daemon/index.js
@@ -6,6 +6,8 @@ var logPath;
 gn.setConfigPath('node_modules/gracenode/scripts/configs/');
 gn.setConfigFiles(['daemon.json']);
 
+gn.exitOnBadOption();
+
 gn.defineOption('--log', 'Enables logging into files in the given path. Example: node daemon start app.js --log=./daemon-logs/', function (path) {
 	logPath = path;
 });